auto rule application

+++ android/app/src/main/java/com/action/email/google/GmailHistoryFetcher.java
@@ -19,6 +19,7 @@ import okhttp3.Request;
 import okhttp3.Response;
 
 import com.action.email.data.MessageAggregateData;
+import com.action.email.data.ProcessRule;
 import com.action.email.realm.model.Message;
 import com.action.email.realm.model.MessageAggregate;
 import com.action.email.realm.service.GmailSyncStateService;
@@ -99,7 +100,7 @@ public class GmailHistoryFetcher {
                                 messageIds.add(messageId);
 
                             }
-                            handleMessageAdded(messageIds);
+                            handleMessageAdded(messageIds );
                         }
 
                         // Handle messageDeleted
@@ -162,7 +163,11 @@ public class GmailHistoryFetcher {
         Log.d(TAG, "Message Added: " + messageIds);
         try {
             ///aggregater is being called from the inside only
-            gmailMessageFetcher.retryBatch(messageIds);
+           List<Message> messages = gmailMessageFetcher.retryBatch(messageIds);
+
+           if(messages != null) {
+               ProcessRule.takeActionOnNewMessages(messages, appContext);
+           }
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
@@ -171,8 +176,11 @@ public class GmailHistoryFetcher {
     private void handleMessageDeleted(String messageId) {
         Log.d(TAG, "Message Deleted: " + messageId);
         Message message = MessageService.getById(messageId);
-        MessageAggregateData.onMessageDeleted(message);
-        MessageService.delete(message);
+        if(message != null) {
+            MessageAggregateData.onMessageDeleted(message);
+            MessageService.delete(message);
+        }
+
 
        // deleteMessageFromRealm(messageId);
     }
@@ -180,16 +188,22 @@ public class GmailHistoryFetcher {
     private void handleLabelAdded(String messageId, String labelId) {
         Log.d(TAG, "Label Added: messageId=" + messageId + " labelId=" + labelId);
         Message message = MessageService.getById(messageId);
-        MessageAggregateData.onLabelAdded(message, labelId);
-        MessageService.addlabel(message, labelId);
+        if(message != null) {
+            MessageAggregateData.onLabelAdded(message, labelId);
+            MessageService.addlabel(message, labelId);
+        }
+
     }
 
     private void handleLabelRemoved(String messageId, String labelId) {
         Log.d(TAG, "Label Removed: messageId=" + messageId + " labelId=" + labelId);
         // Your logic: remove labelId from message labels in Realm
         Message message = MessageService.getById(messageId);
-        MessageAggregateData.onLabelRemoved(message, labelId);
-        MessageService.removeLabel(message, labelId);
+        if(message != null) {
+            MessageAggregateData.onLabelRemoved(message, labelId);
+            MessageService.removeLabel(message, labelId);
+        }
+
     }
